Add exception for mock_snakemake
A common error  when running or debugging a script independetly
is that the "rule is missing". This happens if a conditional rule
is used but not activated in the Snakefile. What is clear in
retrospective was not clear at the moment the error appeared and
took a while to debug.

This commit aims to speed up the users to find the error.